control-service: Allow for jobs with no schedule to be deployed
Currently, if a job is deployed with no schedule_cron,
the deployment will fail. After this change, jobs with
no schedule will be deployed successfully, their cronjobs
will have a default schedule "0 0 30 2 *" - meaning Feb 30,
or never - but the job metadata will have an empty string
for a schedule, so that "vdk list" will not print a schedule
for those jobs (but jobs with their schedule explicitly set
to "0 0 30 2 *" will still have it show).

Testing done: TBD

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>