[vdk-plugins] vdk-control-api-auth: Fix query key type
When parsing the query parameters from the redirect_url in the
authorization code authentication flow, we converted the key of
the parameter to bytes, which was causing TypeErrors due to missing
encoding. This regression was not caught by the unit tests.

This change fixes the regression by using plain strings as keys
and removing any conversions to bytes.

Additionally, unit test was fixed and extended to catch possible
future regressions.

Testing Done: Unit tests

Signed-off-by: Andon Andonov <andonova@vmware.com>