vdk-control-cli: Fix command printed on successful deploy
When successfully deploying a job, vdk-control-cli
prints a command which can be used to check whether
the deployment version was successfully updated.
However, the printed command included the team name
without brackets around it, which can be an issue as
team names can have empty spaces in them, which meant
the command gets parsed wrongly and fails.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>