vdk-control-cli: Adopt new vdk-control-api-auth library
With the release of the vdk-control-api-auth library, we now have a
stand-alone authentication package that is to be used by all Versatile
Data Kit components that need to authenticate against a third-party
service.

This change adopts the new authentication package in vdk-control-cli,
and is now used when issuing `vdk login` commands.

Additionally, authentication logic in vdk-control-cli which is overlapping
with such logic in vdk-control-api-auth is removed.

Testing Done: Installed the new changes locally and ran `vdk login` to verify
that the browser redirecs work as expected. Also, ran a docker container with
the latest changes to verify that authentication is possible even when there
is no access to a browser.

Signed-off-by: Andon Andonov <andonova@vmware.com>