vdk-core: fix buggy (false positive) connection unit test
We have a unit test that checks the raw (or native) connection is being
saved properly - but the test was buggy since they are qual in a sense
they have the same attribute values (as there's only one "is_valid") .
But we care if it's the same instance. So we switch to "is"

Testing Done: After changing == to "is" ran the test and it failed (as
it should) and then fix the test.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>