[vdk-plugins] vdk-control-api-auth: Redefine exceptions
Currently, we have one top-level exception class, `VDKAuthException` which
is invoked every time some error occurs during the authentication process.
This approach works in general, but a downside is that it does not allow for
more granular error handling.

Additionally, if a user wants to integrate their project with the library,
and they want to implement different behaviours based on different errors,
they would need to parse the error message, which may not be considered a
good practice.

This change introduces a new exception hierarchy, which defines child exceptions
to the top-level `VDKAuthException` and facilitates more robust error-handling,
both at the plugin level, as well as by client applications that integrate with
the plugin.

Testing Done: Unit tests

Signed-off-by: Andon Andonov <andonova@vmware.com>