vdk-heartbeat: pipelines-control-service-integration-tests image rebuild
The vdk-heartbeat release documentation required a merge to master, for
triggering the release process. However, the control-service CI/CD that
depends on vdk-heartbeat, does not consider new vdk-heartbeat versions.

Added a comment to control-service `.gitlab-ci.yml` to trigger an IT
image update. Documented a workaround in `README.md` until automated.

Testing Done: will monitor the pipelines for image rebuild

Signed-off-by: ikoleva <ikoleva@vmware.com>