[vdk-plugins] airflow-provider-vdk: Adopt auth plugin
After the release of the `vdk-control-api-auth` plugin, which acts as
the de-facto authentication package for Versatile Data Kit, we no longer
need to depend on `vdk-control-cli` and all its unnecessary dependencies.

With this change, we adopt the `vdk-control-api-auth` package and its
authentication capabilities, and drop any dependency on `vdk-control-cli`.

Testing Done: Unit tests.

Signed-off-by: Andon Andonov <andonova@vmware.com>