versatile-data-kit: Airflow example
This change introduces an example where the user deploys
three jobs: two of these ingest data from predefined
sources into a target database and the third aggregates
parts of the data. The example also guides the user to
instantiate an Airflow DAG which runs these jobs in a
dependent manner, meaning that the two ingest jobs must
pass successfully before the third job is ran.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>