control-service: Atomic job cancellation
Currently, there is a possibility for a job to be set to be cancelled,
and have it complete before the cancellation can complete. This
leads to a 500 error with no explanation.

This change introduces an additional check that the cancellation does
not raise a NullPointerException, and if it does, it logs an appropriate
message and raises and appropriate Exception.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>