vdk-core: Fix destination_table referenced early
In the IngesterBase class, we currently pass `destination_table` to the
call to `_queue_payload_for_posting` in cases when the objects queue is empty.
However, at this point, the destination_table is not defined, which causes an
UnboundLocalError to be raised.

This change fixes the issue by replacing `destination_table` with `current_destination_table`
(set to None by default) which is defined before the call to `_queue_payload_for_posting`.

Testing Done: All existing tests pass

Signed-off-by: Andon Andonov <andonova@vmware.com>