vdk-core: Add unit test for destination_table in empty queue
As part of https://github.com/vmware/versatile-data-kit/pull/864, a change was
made that fixes an issue with destination_table referenced before assignment in
case the `_objects_queue` of the `IngesterBase` class is empty. However, the change
was merged without any tests.

This change introduces a unit test that verifies that there are no regressions.

Testing Done: Unit test introduced.

Signed-off-by: Andon Andonov <andonova@vmware.com>