control-service: improve api message on oom job execution errors (#861)
why: Recently users of control-service have been complaining that jobs
which fail with User Error and have failed with an OOM error didn't
have a sufficiently clear enough message that classifies such types of errors.

what: Changed internal implementation so that when a data job fails due
to K8S OOM errors the message stored in the database contains a hint as
to the cause of failure - out of memory issues. We now explicitly check the
container termination message and in case of User Errors with an OOM error
we set an appropriate message.

testing: Refactored existing tests. Added new unit tests.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com