vdk-core: detect non vdk_ prefixed environment values for config
Currently user need to set VDK_ + config name as environment variable so
VDK  can recognize it. The idea was to avoid conflicts with similarly
named env variables. But this proved to be very error prone where users
tend to forget or get confused on where vdk_ prefix is needed and where
it is not.

This is solved by allowing non-prefixed env variable to be set.

For for configuration "team"

Now if "VDK_TEAM" is set as environment variable is set it is used (as
before)
If it is not set but "TEAM" is set as env variable, it is now used.

Prefixed variables do have higher priority.

Testing Done: unit test

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>