[vdk-plugins] airflow-provider-vdk: Add hidden fields to VDK Connection (#883)
In the Airflow UI, users can create new connections manually from the Admin section. When
done this way a web form is shown, where they can fill in host, port, username, password,
etc. needed to establish a new connection to third-party services.

The fields shown in the web form can be customised by overwriting the `get_ui_field_behaviour()`
method from the **BaseHook**. In the vdk airflow provider, we do some field relabeling to make
the user experience more consistent with our project.

This, however, caused some issues, as apperantly the dictionary where the changes to the connection
fields' behaviour is specified, needs to have a `hidden_fields` key/value pair.

This change introduces a `hidden_fields` key/value pair, and we use it to hide the schema field, which
is not really used by vdk.

Testing Done: Tested with local airflow instance and veryfied no errors in logs and correct web form
is shown to users.

Signed-off-by: Andon Andonov <andonova@vmware.com>