vdk-core: vdk_exception hook exit code fix
In case of StandaloneDataJob, the vdk_exception hook was always
returning exit code 1, no matter if exception was successfully handled
or not. Also, in case of empty job, vdk_exception was not triggered.

Fixed the StandaloneDataJob return status to 0 in case exception was
successfully handled. Fixed the case with ClickException not propagating
to vdk_exception hook.

Testing Done: run tests locally; did verify manually that an empty job
does evaluate vdk_exception hook, and in case exception is successfully
handled then the data job exit code is 0.

Signed-off-by: ikoleva <ikoleva@vmware.com>