control-service: fix vdk-server startup issues (#908)
what: Added default values for env variable that was causing the control-service to crash on startup, because even though the authentication/authorization is disabled the variables are used in a bean (constructor) initialization and it crashes on spring boot initialization (before it enters our code). The env variable is found in some of our application.properties files, but it varies between prod, testing etc (not all have default values) and there are comments that say setting it is mandatory.

why: the vdk-server's vdk server --install command was failing.

testing: Tested locally with pip install -e option of a locally editable vdk-server plugin. Server was failing to start without the variable. After the change - server starts up normally.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com