vdk-core: execution result missing exception and blamee fix (#938)
* vdk-core: execution result missing exception and blamee fix

The execution result was sometimes and sometimes not provisioned with
the exception in case such is known. It was also lacking information
about blamee, yet interfacing unclassified (raw) exceptions.
Therefore, the information about error classification was not present
in ExecutionResult, only in StepResult (design flaw).
Also, data job execution summary was not referring to the exception
dynamically (via getter), that was prone to non-initialized exception.

Improved the `ExecutionResult` design by elaborating on blamee alongside
a raw-or-not-raw exception. Fixed the rest of the ExecutionResult
initialization implementation so that in all the cases the exception is
known, then it is populated.
Refined the data job execution summary by adding exception_name
and blamee in case of job failure.

Testing Done: Test-fixed expectation about a non-populated exception
in ExecutionResult, and added blamee check.
Updated tests to using `get_exception_to_raise()` in favour of the
deprecated `get_exception()`.

Signed-off-by: ikoleva <ikoleva@vmware.com>