vdk-core: Fix issue with serializing Decimal values during payload check
As part of the ingestion process, the payload passes a
check to verify it is JSON serializable. However,
`json.dumps` does not natively support serializing
Decimal objects, which leads to the check failing when
it should not. This change fixes this problem by amending
the JsonEncoder object to convert Decimals to floats
during the check.
Note that this does not affect the actual ingestion of
Decimal objects, merely the serialization check.

Testing done: TBD

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>