vdk-core: avoid circular references in print results
If there's a circular reference in the data printed by
execution_results, json.dumps will fail.
So we are failing back to pprint which can handle them.

pprint would sort the keys for us which leads to more confusing message
(e.g the name of the stpe is not first for each step, etc.) That's why
we are not using it by default.

Sorting can be disabled but only from 3.8+

Testing Done: extended unit tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>