vdk-properties-fs: new plugin for local FS properties storage
We need a simplistic plugin, that allows a developer or presenter
to quickly store secrets on the local FS during a dev/demo session,
that does not require a prerequisite of the entire Control Service
installed and running.

Added a FS properties client, that uses a customizable dir and filename
on the local file system to store the properties key-values. It does
support multiple teams and data jobs, the properties of those are placed
within a dedicated section for keys deduplication.

Testing Done: added `test_fs_properties_plugin.py`, that covers multiple
data jobs using matching keys, and default/customized directory and
filename for file storage on the local FS.

Signed-off-by: ivakoleva <iva.koleva@clearcode.bg>