vdk-core: Add log before query result fetch
Occasionally, vdk may indicate that a query execution took 10 or 20
minutes, even though the DB back shows that the query was actually
executes in seconds. This results in the DB backend issuing a timeout
error and a job failure. Because it happens sporadically, it is hard
to reproduce and debug locally.

This change adds a log message just before the result of a query is
fetched, and should help better understand at which step the query
execution hanged.

Testing done: Documentation change.

Signed-off-by: Andon Andonov <andonova@vmware.com>