vdk-core: errors occurred and the state (handled or not) context missing
Re submit the change -
https://github.com/vmware/versatile-data-kit/pull/1182 - again. The only
difference is that I renamed _get_eception_message to
get_exception_message as not to break backwords compatability. I will
merge this after

--

User code may trigger a SDK feature in a way that raises a user error.
In such a use case, the error could be handled by the cautious user, in
runtime. The VDK platform should be aware if user errors are handled
in runtime or not, so that affects the data job termination status.

The old errors.BLAMEES (experimental data structure for
keeping track of responsibles for errors priorly logged) should be
refactored:
* a resolvable context abstraction introduced
* it is memory-managed (singleton, reasoning documented)
* populated by the 3 user-facing `errors.log_*()` methods - in favour of
a method designed to populate ErrorMessage entries
* keeps track if the actual exception, along with all properties like
who is the actual respobsible persona (user or platform, that was
free-form str before), resolved by a data job step or not, etc.

Testing Done: added a functional test covering the
 `simple-query-failed-handled` scenario, added singleton verification
 for the resolvable context; ci/cd

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>

vdk-core: test

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>