versatile-data-kit: required approving reviewers update
On the latest OSS meeting, we have decided to introduce the 2 code
reviewers bar, so that a change should meet in order to be merged.
This improves the code quality, and team context of the changes made.

I have updated the CONTRIBUTING.md guide, yet looking for accessing the
project settings, so we can update the Required approving reviewers
https://github.blog/2018-03-23-require-multiple-reviewers/

Testing Done: n/a

Signed-off-by: ivakoleva <iva.koleva@clearcode.bg>