vdk-pipeline-control-service: upgrade java client k8s version (#1216)
# Why?
I upgraded java client k8s version when trying to fix an issue related to connecting to the k8s cluster. 
This was ultimately not the fix that was needed. But considering the work is done I think it make sense to merge it regardless. 

# What?
Upgrade the version in gradle and make a few small code changes to support the new apis. 

# How has this been tested?
Unit tests, integration tests but no new tests have been added

# What type of change are you making?
Upgrade the version in gradle and make a few small code changes to support the new apis. 


Signed-off-by: murphp15 <murphp15@tcd.ie>