vdk-plugins: run tests on release of vdk-core (#1210)
Changes in vdk-core can break some plugins. We need some functionality
that enables us to catch such regression quicker.

The proposoed change is to introduce in .plugin-common.yml new generic
"job" template `.build-plugin-on-vdk-core-release` which can be used by
plugins CI.

As it should not be necessary to run all plugin tests, we can select
some more important

Testing Done: Setup CI to trigger the enw job and verified it's using
vdk-core dev version for the tests.