VDK conferences: Add a folder in the main repo to store the "Welcome pages" of the conferences we attend (#1234)
Having a folder inside VDK repo gives visibility to the community on which conferences did we present. It also gives a direct link between the main VDK repo and the repo we use for storing VDK examples/workshops used at the conferences themselves (https://github.com/versatile-data-kit-demo). Also, people at the conference could directly star the VDK repo instead of the workshops repo.

Added a folder "conferences" with subfolders for AMLD and OpenFest at which we already presented/will present. Inside each conf. subfolder, there's a README which is the "home/welcome" page shown on the conference and any other files that might be relevant.

Cosmetic change/folder structure