vdk-impala: Refactor query classifier for data lineage
# Why?
Analysing the SQL query is not so trivial task, as there are number of variations of the DDL, DML and DDL with DML combinations. Current avoidance of single line comments does not work properly and skips queries that are eligible for lineage collection.

# What?
Added sqlparse library in order to externalise comment-cleaning task, now it leaves us only with the query and there isn't any need to analyse whole query. What we do now is we check if query starts with blacklisted keyword, if so - query is not eligible for lineage. Although some keywords blacklisted might contain lineage data, like CREATE statement, when table is created from another with SELECT statement. In this case blacklisted word needs to be whitelisted in some conditions.

As sql statements are complex, we need to get which queries definitely has data lineage, instead of searching which query have for sure and for sure not data lineage. False positives queries for lineage, ie queries that does not have lineage information provided from the profile, but are marked as having one, will not break the logic at all, instead we optimize load on the impala server with a simple approach

# How has this been tested?
Unit and integration tests, also with production-grade queries

# What type of change are you making?
Bug-fixing

Signed-off-by: Plamen Kostov <pkostov@vmware.com>