vdk-meta-jobs: Initial implementation
This is an pre-Alpha implementation of Meta Jobs functioanlity

One crucial feature which has been missing from VDK is the ability to
express workload (data job) interdependencies, meaning that a certain
Data Job wil only trigger upon the successful completion (or other
criteria) of one or more other Data Jobs.

An integration with Apache Airflow was developed to address this,
however this comes with its downsides: namely that teams would have to
manage their own instance of Airflow, or pay for an externally managed
service.

The aim of this submission is to build a plugin for Versatile Data Kit
which extends its API with an additional feature which allows users to
trigger so called Meta Jobs - jobs which trigger one or more other jobs,
wait for their completion, and then trigger another set of jobs until
either the entire job pipeline has succeeded, or at least one of the
jobs has failed.

https://github.com/vmware/versatile-data-kit/issues/1243

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>