[vdk-plugins] vdk-ingest-http: Move data conversion above size calc
Currently, the memory size of the payload to be ingested is calculated before it is
converted to string. This is fine for most cases, however, it gets a bit tricky for
more complex objects, due to how memory size is calculated (see
https://stackoverflow.com/questions/63204377/memory-size-of-list-python for reference).

This change moves the conversion of the payload to string before the memory size is calculated,
so we avoid posibility to return misleading IngestionResult data.

Testing Done: Cosmetic change, all tests pass.

Signed-off-by: Andon Andonov <andonova@vmware.com>