control-service: Dynamic python site-packages directory detection
The python site-packages directory is currently hardcoded in
the data job container command. However,
this works only for data jobs based on Python 3.7.

This change aims to make the data job cronjob template agnostic
for the python version by replacing the hardcoded python
site-packages directory with a python command that returns it.

Tested locally against the kind cluster.

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com