vdk-core: Allow for modification of dynamic params
The ingestion functionality in vdk-core allows for the creation of pre-processor
plugins that modify the passed payload. This works for most use-cases, but not
when an operator of Versatile Data Kit needs to modify or otherwise ammend the
dynamically set parameters like target, destination_table, etc., to comply with
specific rules.

This change adds functionality to allow for the modification of dynamically set
parameters like target by pre-processor plugins. Once modified, the parameters are
to be placed in a dictionary, which is then to be added to the metadata using the
`updated_dynamic_params` key.

Testing Done: Unit test

Signed-off-by: Andon Andonov <andonova@vmware.com>