control-service: refactor db query to mitigate race condition (#1269)
what: Refactored how the watch task updates missing kubernetes job executions. It should now update the statuses directly against the database, instead of loading entities in memory, updating them (in memory) and saving to the database.

why: Users reported that when cancelling executions, the cancelled execution's status is visible as Succeeded, even tough the API returned success and the status was set to Cancelled. This is caused by the watch task which updates executions in bulk and in memory before saving them to the database.

testing: Added unit tests that cover the new repository query.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com