vdk-postgres: add ingest plugin
When we were preparing for DSC conference we realized that we are
missing simple ingest plugin.

This is really copy pasting the greenplum ingestion. where I changed
"greenplum" to "postgres"

I will add this now since we need it . But I think it's time to refactor
the db plugins - https://refactoring.guru/refactoring/when

A lot of the code is basically the same with names changed. ... So I
will open a PR for that.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>