vdk-logging-json: Introduce configuration
Currently, the vdk-logging-json plugin will format logs
to JSON as long as it is installed. This is suboptimal,
as it would be better for this to be configurable in the
case where a user might have multiple log formatting
plugins installed.

Additionally, one detail which should be mentioned
is that the initial VDK logs are printed before the
config is initialized. This means that to format these
logs to JSON, we check for the `VDK_LOGGING_FORMAT`
env variable instead of accessing the configuration
through the config object. In the case where the
logging format is set through a configuration plugin
instead of an environment variable, the first few
logs will not be formatted correctly. I do not believe
there is a good solution to this problem which wouldn't
require large architectural changes to VDK.

Testing done: tested all cases locally

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>