control-service: builder images load secrets from k8s (#1358)
Why
I wrote this PR with the intent of being able to pass extra docker creds to to the image builder so that we would be able to pull a private base image from a docker repo.
The builder now contains login details for 2 registries. The registry it pushes to and the registry it pulls base image from.

However this PR also introduces the concept of builder secrets being sourced from k8s which leads to much cleaner separation between the control plane and the builder image. I think this will lead to much better support for custom build images in the future.

I'm not sure where to add the documentation about the builder-secrets. So I created a ticket to follow up on this later.
#1357

How was this been tested
tested extensively locally.
I added a test to check that the base image can be pulled from a private repo but in fact it is pulling from the same repo as it is pushing to and so it is could let bugs through related to the auth not working for the base image repo.

Signed-off-by: murphp15 <murphp15@tcd.ie>

* Google Java Format

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: github-actions <>