control-service: run release  test on dependency version change
In case of change in versions-of-external-dependencies.gradle we are directly making a new release without triggering release acceptance test which may result in broken release. It's better to make sure the release is tested. 

Also removed job-builder changes from trigger deployment and release test for now as generally default job-builder version is pinned so when changed it would trigger new test. We need to devise a way to test all differnt job-builders anyhow. 

Finally refactored slighlty the yaml anchors to re-use previously defined ones.

Testing Done: this PR