examples: clarify README sample anonymize plugin (#1394)
The sample README used in a tutorial may be misleading that it's a real
plugin (since it looks like a README of a real plugin). This may confuse
people and try to use it. So we are clarifying it here to avoid
mistakes.

This was flagged by VMware security team as a potential dependecy
confusion attack.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>