control-service: Reset termination status when job is disabled
Currently, when a data job is deleted, we reset its termination status, so as to avoid
issues with monitoring services reporting incorrect status. This, however, is not done
in case a data job is disabled, even though the job is no longer active and its termination
status should not matter.

This change modifies the DeplymentService, so that if a data job is disabled, the termination
status is reset.

Testing Done: Added test.

Signed-off-by: Andon Andonov <andonova@vmware.com>