versatile-data-kit: VEP-1418 on VDK connectors
VDK users need to be able to enlist the VDK database connectors
available. Currently, the happy path is for the users to come up with
the idea of manually browsing the `vdk-plugins/*` documentation and
sources, then decide by their own what seems to be a database connector.
By automating this lookup, we provide an instrument for connectors
discovery, saving users time and efforts, and preventing human errors.

Added an issue https://github.com/vmware/versatile-data-kit/issues/1418
and initiated a VEP specification.

Testing Done: N/A

Signed-off-by: ivakoleva <iva.koleva@clearcode.bg>